Make the server package build and test on Windows
Split the disk-status check into per-platform files so the package
compiles everywhere: the existing Linux statfs implementation, a Windows
one backed by GetDiskFreeSpaceEx, and a neutral fallback for other
platforms. golang.org/x/sys becomes a direct dependency.

Two existing tests that cannot work on Windows now skip there instead of
failing: the cjs-module-lexer binary install (prebuilt binaries are only
published for darwin/linux) and the esbuild plugin path resolution case.
This lets contributors run go test ./server/ locally on Windows.
