+++ go.mod
@@ -10,10 +10,8 @@ require (
 	github.com/ije/gox v0.10.4
 	golang.org/x/crypto v0.57.0
 	golang.org/x/net v0.59.0
+	golang.org/x/sys v0.48.0
 	golang.org/x/term v0.46.0
 )
 
-require (
-	golang.org/x/sys v0.48.0 // indirect
-	golang.org/x/text v0.42.0 // indirect
-)
+require golang.org/x/text v0.42.0 // indirect
