+++ server/disk.go
@@ -3,7 +3,6 @@ package server
 import (
 	"os"
 	"sync"
-	"syscall"
 
 	"github.com/ije/gox/log"
 )
@@ -48,20 +47,4 @@ func purgeNPMCacheWhenDiskIsLowOrFull(npmrc *NpmRC, logger *log.Logger) {
 	if err := os.RemoveAll(oldDir); err != nil {
 		logger.Errorf("failed to remove npm cache directory %s: %v", oldDir, err)
 	}
-}
-
-func checkDiskStatus() DiskStatus {
-	var stat syscall.Statfs_t
-	err := syscall.Statfs(config.WorkDir, &stat)
-	if err == nil {
-		avail := stat.Bavail * uint64(stat.Bsize)
-		if avail < 100*MB {
-			return DiskStatusFull
-		} else if avail < 1024*MB {
-			return DiskStatusLow
-		}
-	} else {
-		return DiskStatusError
-	}
-	return DiskStatusOk
 }
